Remove override from combat dummies
