+++ game/src/main/kotlin/content/entity/player/combat/Attack.kts
@@ -10,8 +10,15 @@ import content.skill.melee.weapon.attackRange
 import content.entity.combat.combatPrepare
 import content.entity.player.dialogue.type.statement
 import content.skill.magic.spell.spell
+import content.skill.melee.weapon.fightStyle
+import net.pearx.kasechange.toTitleCase
+import world.gregs.voidps.engine.client.message
+import world.gregs.voidps.engine.entity.character.mode.interact.TargetInteraction
+import world.gregs.voidps.engine.entity.character.npc.NPC
+import world.gregs.voidps.engine.entity.character.npc.NPCOption
 import world.gregs.voidps.engine.entity.character.npc.npcApproach
 import world.gregs.voidps.engine.entity.character.npc.npcApproachNPC
+import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.equip.equipped
 import world.gregs.voidps.network.login.protocol.visual.update.player.EquipSlot
 
@@ -19,6 +26,9 @@ npcApproach("Attack") {
     if (player.equipped(EquipSlot.Weapon).id.endsWith("_greegree")) {
         statement("You cannot attack as a monkey.")
         cancel()
+        return@npcApproach
+    }
+    if (target.id.endsWith("_dummy") && !handleCombatDummies()) {
         return@npcApproach
     }
     if (character.attackRange != 1) {
@@ -50,6 +60,10 @@ characterApproachPlayer("Attack") {
 itemOnNPCApproach(id = "*_spellbook") {
     approachRange(8, update = false)
     player.spell = component
+    if (target.id.endsWith("_dummy") && !handleCombatDummies()) {
+        player.clear("spell")
+        return@itemOnNPCApproach
+    }
     player["attack_speed"] = 5
     player["one_time"] = true
     player.attackRange = 8
@@ -71,4 +85,16 @@ combatPrepare { player ->
 fun combatInteraction(character: Character, target: Character) {
     val interact = character.mode as? Interact ?: return
     interact.updateInteraction(CombatInteraction(character, target))
+}
+
+suspend fun TargetInteraction<Player, NPC>.handleCombatDummies(): Boolean {
+    val type = target.id.removeSuffix("_dummy")
+    if (player.fightStyle == type) {
+        return true
+    }
+    player.message("You can only use ${type.toTitleCase()} against this dummy.")
+    approachRange(10, false)
+    player.mode = EmptyMode
+    cancel()
+    return false
 }
